+++ src/Command/Redis/BITOP.php
@@ -12,6 +12,7 @@
 
 namespace Predis\Command\Redis;
 
+use InvalidArgumentException;
 use Predis\Command\PrefixableCommand as RedisCommand;
 
 /**
@@ -19,6 +20,8 @@ use Predis\Command\PrefixableCommand as RedisCommand;
  */
 class BITOP extends RedisCommand
 {
+    private const VALID_OPERATIONS = ['AND', 'OR', 'XOR', 'NOT', 'DIFF', 'DIFF1', 'ANDOR', 'ONE'];
+
     /**
      * {@inheritdoc}
      */
@@ -36,6 +39,13 @@ class BITOP extends RedisCommand
             [$operation, $destination] = $arguments;
             $arguments = $arguments[2];
             array_unshift($arguments, $operation, $destination);
+        }
+
+        if (!empty($arguments)) {
+            $operation = strtoupper($arguments[0]);
+            if (!in_array($operation, self::VALID_OPERATIONS, false)) {
+                throw new InvalidArgumentException('BITOP operation must be one of: AND, OR, XOR, NOT, DIFF, DIFF1, ANDOR, ONE');
+            }
         }
 
         parent::setArguments($arguments);